requirements.txt
telebot
mysql-connector
python-dotenv
flask
requests
redis